chore: rename metadb/component.go -> component/metadb.go
